+++ adapters/src/repositories/__init__.py
@@ -1,0 +1,4 @@
+from .email_sender import SMTPEmailSenderRepository
+from .memory import (MemoryAppointmentRepository, MemoryEmailSenderRepository,
+                     MemoryProductRepository)
+from .sql import SQLAppointmentRepository, SQLProductRepository
