+++ adapters/src/repositories/email_sender/__init__.py
@@ -1,0 +1,1 @@
+from .email_sender_adapter import SMTPEmailSenderRepository
