+++ adapters/src/repositories/memory/__init__.py
@@ -1,0 +1,3 @@
+from .memory_appointment_repository import MemoryAppointmentRepository
+from .memory_email_sender_repository import MemoryEmailSenderRepository
+from .memory_product_repository import MemoryProductRepository
