+++ adapters/src/repositories/sql/__init__.py
@@ -1,0 +1,4 @@
+from .config_db.session_manager import Session
+from .config_db.sql_connection import SQLConnection
+from .sql_appointment_adapter import SQLAppointmentRepository
+from .sql_product_adapter import SQLProductRepository
