+++ adapters/src/repositories/sql/tables/__init__.py
@@ -1,0 +1,2 @@
+from .appointment import AppointmentRecord
+from .product import ProductRecord
